requirements.txt
Django==4.2.7
psycopg2-binary==2.9.11
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
requests==2.31.0
